+++ packages/arco-vue-docs/locale/en-us.js
@@ -92,6 +92,9 @@ export default {
     treeSelect: 'TreeSelect',
     overflow: 'OverflowList',
     scrollbar: 'Scrollbar',
+    watermark: 'Watermark',
+    verificationCode: 'VerificationCode',
+    colorPicker: 'ColorPicker',
   },
   footer: {
     design: 'Design',